FEAT: made extension finder more generic
