+++ src/io/readers/spectrum_reader/minitdf.rs
@@ -1,7 +1,4 @@
-use std::{
-    fs,
-    path::{Path, PathBuf},
-};
+use std::path::{Path, PathBuf};
 
 use crate::{
     io::readers::{
@@ -14,6 +11,7 @@ use crate::{
         PrecursorReader,
     },
     ms_data::Spectrum,
+    utils::find_extension,
 };
 
 use super::SpectrumReaderTrait;
@@ -85,21 +83,4 @@ impl SpectrumReaderTrait for MiniTDFSpectrumReader {
     fn get_path(&self) -> PathBuf {
         self.path.clone()
     }
-}
-
-fn find_extension(path: impl AsRef<Path>, extension: &str) -> Option<PathBuf> {
-    let extension_lower = extension.to_lowercase();
-    for entry in fs::read_dir(&path).ok()? {
-        if let Ok(entry) = entry {
-            let file_path = entry.path();
-            if let Some(file_name) =
-                file_path.file_name().and_then(|name| name.to_str())
-            {
-                if file_name.to_lowercase().ends_with(&extension_lower) {
-                    return Some(file_path);
-                }
-            }
-        }
-    }
-    None
 }
